refactor: auth service

+++ packages/apps/reputation-oracle/server/src/modules/auth/auth.service.ts
@@ -1,34 +1,27 @@
-import {
-  KVStoreClient,
-  KVStoreKeys,
-  KVStoreUtils,
-  Role,
-} from '@human-protocol/sdk';
+import { KVStoreKeys, KVStoreUtils, Role } from '@human-protocol/sdk';
 import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 
-import {
-  SiteKeyType,
-  UserStatus,
-  UserRole,
-  UserEntity,
-  UserRepository,
-  UserService,
-  OperatorStatus,
-  type Web2UserEntity,
-  type OperatorUserEntity,
-} from '../user';
-import { TokenEntity, TokenType } from './token.entity';
-import { TokenRepository } from './token.repository';
-import { verifySignature } from '../../utils/web3';
-import { Web3Service } from '../web3/web3.service';
 import { SignatureType } from '../../common/enums/web3';
-import { prepareSignatureBody } from '../../utils/web3';
-import * as securityUtils from '../../utils/security';
 import { AuthConfigService } from '../../config/auth-config.service';
 import { NDAConfigService } from '../../config/nda-config.service';
 import { ServerConfigService } from '../../config/server-config.service';
 import { Web3ConfigService } from '../../config/web3-config.service';
+import * as web3Utils from '../../utils/web3';
+import * as securityUtils from '../../utils/security';
+import { EmailAction } from '../email/constants';
+import { EmailService } from '../email/email.service';
+import logger from '../../logger';
+import {
+  SiteKeyRepository,
+  SiteKeyType,
+  UserStatus,
+  UserEntity,
+  UserRepository,
+  UserService,
+  type OperatorUserEntity,
+  type Web2UserEntity,
+} from '../user';
 import {
   AuthError,
   AuthErrorMessage,
@@ -51,14 +44,14 @@ import {
   VerifyEmailDto,
   ResendVerificationEmailDto,
 } from './dto';
-
-import { EmailService } from '../email/email.service';
-import { EmailAction } from '../email/constants';
+import { TokenEntity, TokenType } from './token.entity';
+import { TokenRepository } from './token.repository';
 
 @Injectable()
 export class AuthService {
-  private readonly salt: string;
-
+  private readonly logger = logger.child({
+    context: AuthService.name,
+  });
   constructor(
     private readonly jwtService: JwtService,
     private readonly userService: UserService,
@@ -68,27 +61,11 @@ export class AuthService {
     private readonly ndaConfigService: NDAConfigService,
     private readonly web3ConfigService: Web3ConfigService,
     private readonly emailService: EmailService,
-    private readonly web3Service: Web3Service,
     private readonly userRepository: UserRepository,
+    private readonly siteKeyRepository: SiteKeyRepository,
   ) {}
 
-  public async signin({
-    email,
-    password,
-  }: Web2SignInDto): Promise<SuccessAuthDto> {
-    const userEntity = await this.userService.findWeb2UserByEmail(email);
-    if (!userEntity) {
-      throw new AuthError(AuthErrorMessage.INVALID_CREDENTIALS);
-    }
-
-    if (!securityUtils.comparePasswordWithHash(password, userEntity.password)) {
-      throw new AuthError(AuthErrorMessage.INVALID_CREDENTIALS);
-    }
-
-    return this.auth(userEntity);
-  }
-
-  public async signup(data: Web2SignUpDto): Promise<void> {
+  async signup(data: Web2SignUpDto): Promise<void> {
     const storedUser = await this.userRepository.findOneByEmail(data.email);
     if (storedUser) {
       throw new DuplicatedUserEmailError(data.email);
@@ -109,7 +86,216 @@ export class AuthService {
     });
   }
 
-  public async refresh(data: RefreshDto): Promise<SuccessAuthDto> {
+  async web3Signup(
+    data: Web3SignUpDto,
+  ): Promise<{ accessToken: string; refreshToken: string }> {
+    const user = await this.userRepository.findOneByAddress(data.address);
+    if (user) {
+      throw new DuplicatedUserAddressError(data.address);
+    }
+
+    const preSignUpData = web3Utils.prepareSignatureBody({
+      from: data.address,
+      to: this.web3ConfigService.operatorAddress,
+      contents: SignatureType.SIGNUP,
+    });
+
+    const isValidSignature = web3Utils.verifySignature(
+      preSignUpData,
+      data.signature,
+      [data.address],
+    );
+
+    if (!isValidSignature) {
+      throw new AuthError(AuthErrorMessage.INVALID_WEB3_SIGNATURE);
+    }
+
+    const chainId = this.web3ConfigService.reputationNetworkChainId;
+    const role = await KVStoreUtils.get(
+      chainId,
+      data.address,
+      KVStoreKeys.role,
+    );
+
+    // We need to exclude ReputationOracle role
+    const isValidRole = [
+      Role.JobLauncher,
+      Role.ExchangeOracle,
+      Role.RecordingOracle,
+    ].includes(role);
+
+    if (!isValidRole) {
+      throw new InvalidOperatorRoleError(role);
+    }
+
+    const fee = await KVStoreUtils.get(chainId, data.address, KVStoreKeys.fee);
+    if (!fee) {
+      throw new InvalidOperatorFeeError(fee);
+    }
+
+    const url = await KVStoreUtils.get(chainId, data.address, KVStoreKeys.url);
+    if (!url) {
+      throw new InvalidOperatorUrlError(url);
+    }
+
+    const jobTypes = await KVStoreUtils.get(
+      chainId,
+      data.address,
+      KVStoreKeys.jobTypes,
+    );
+    if (!jobTypes) {
+      throw new InvalidOperatorJobTypesError(jobTypes);
+    }
+
+    const userEntity = await this.userService.createOperatorUser(data.address);
+
+    return this.web3Auth(userEntity);
+  }
+
+  async signin({
+    email,
+    password,
+  }: Web2SignInDto): Promise<{ accessToken: string; refreshToken: string }> {
+    const userEntity = await this.userService.findWeb2UserByEmail(email);
+    if (!userEntity) {
+      throw new AuthError(AuthErrorMessage.INVALID_CREDENTIALS);
+    }
+
+    if (!securityUtils.comparePasswordWithHash(password, userEntity.password)) {
+      throw new AuthError(AuthErrorMessage.INVALID_CREDENTIALS);
+    }
+
+    return this.auth(userEntity);
+  }
+
+  async web3Signin(data: Web3SignInDto): Promise<SuccessAuthDto> {
+    const userEntity = await this.userService.findOperatorUser(data.address);
+
+    if (!userEntity) {
+      throw new AuthError(AuthErrorMessage.INVALID_ADDRESS);
+    }
+
+    const preSigninData = web3Utils.prepareSignatureBody({
+      from: data.address,
+      to: this.web3ConfigService.operatorAddress,
+      contents: SignatureType.SIGNIN,
+      nonce: userEntity.nonce,
+    });
+    const isValidSignature = web3Utils.verifySignature(
+      preSigninData,
+      data.signature,
+      [data.address],
+    );
+
+    if (!isValidSignature) {
+      throw new AuthError(AuthErrorMessage.INVALID_WEB3_SIGNATURE);
+    }
+
+    const nonce = web3Utils.generateNonce();
+    await this.userRepository.updateOneById(userEntity.id, { nonce });
+
+    return this.web3Auth(userEntity);
+  }
+
+  async auth(
+    userEntity: Web2UserEntity | OperatorUserEntity | UserEntity,
+  ): Promise<{ accessToken: string; refreshToken: string }> {
+    let siteKey: string | undefined;
+    const hCaptchaSiteKey = await this.siteKeyRepository.findByUserAndType(
+      userEntity.id,
+      SiteKeyType.HCAPTCHA,
+    );
+    if (hCaptchaSiteKey && hCaptchaSiteKey.length > 0) {
+      // TODO: Revisit
+      // We know for sure that only one hcaptcha sitekey might exist
+      siteKey = hCaptchaSiteKey[0].siteKey;
+    }
+
+    // TODO: DRY
+    const jwtPayload = {
+      email: userEntity.email,
+      status: userEntity.status,
+      user_id: userEntity.id,
+      wallet_address: userEntity.evmAddress,
+      role: userEntity.role,
+      kyc_status: userEntity.kyc?.status,
+      nda_signed:
+        userEntity.ndaSignedUrl === this.ndaConfigService.latestNdaUrl,
+      reputation_network: this.web3ConfigService.operatorAddress,
+      qualifications: userEntity.userQualifications
+        ? userEntity.userQualifications.map(
+            (userQualification) => userQualification.qualification.reference,
+          )
+        : [],
+      site_key: siteKey,
+    };
+
+    const refreshTokenEntity =
+      await this.tokenRepository.findOneByUserIdAndType(
+        userEntity.id,
+        TokenType.REFRESH,
+      );
+
+    if (refreshTokenEntity) {
+      await this.tokenRepository.deleteOne(refreshTokenEntity);
+    }
+
+    const newRefreshTokenEntity = new TokenEntity();
+    newRefreshTokenEntity.userId = userEntity.id;
+    newRefreshTokenEntity.type = TokenType.REFRESH;
+    const date = new Date();
+    newRefreshTokenEntity.expiresAt = new Date(
+      date.getTime() + this.authConfigService.refreshTokenExpiresIn * 1000,
+    );
+
+    await this.tokenRepository.createUnique(newRefreshTokenEntity);
+
+    const accessToken = await this.jwtService.signAsync(jwtPayload, {
+      expiresIn: this.authConfigService.accessTokenExpiresIn,
+    });
+
+    return { accessToken, refreshToken: newRefreshTokenEntity.uuid };
+  }
+
+  async web3Auth(
+    userEntity: OperatorUserEntity | UserEntity,
+  ): Promise<{ accessToken: string; refreshToken: string }> {
+    // TODO: DRY
+    const jwtPayload = {
+      status: userEntity.status,
+      user_id: userEntity.id,
+      wallet_address: userEntity.evmAddress,
+      role: userEntity.role,
+      reputation_network: this.web3ConfigService.operatorAddress,
+    };
+    const refreshTokenEntity =
+      await this.tokenRepository.findOneByUserIdAndType(
+        userEntity.id,
+        TokenType.REFRESH,
+      );
+
+    if (refreshTokenEntity) {
+      await this.tokenRepository.deleteOne(refreshTokenEntity);
+    }
+
+    const newRefreshTokenEntity = new TokenEntity();
+    newRefreshTokenEntity.userId = userEntity.id;
+    newRefreshTokenEntity.type = TokenType.REFRESH;
+    const date = new Date();
+    newRefreshTokenEntity.expiresAt = new Date(
+      date.getTime() + this.authConfigService.refreshTokenExpiresIn * 1000,
+    );
+
+    await this.tokenRepository.createUnique(newRefreshTokenEntity);
+
+    const accessToken = await this.jwtService.signAsync(jwtPayload, {
+      expiresIn: this.authConfigService.accessTokenExpiresIn,
+    });
+
+    return { accessToken, refreshToken: newRefreshTokenEntity.uuid };
+  }
+
+  async refresh(data: RefreshDto): Promise<SuccessAuthDto> {
     const tokenEntity = await this.tokenRepository.findOneByUuidAndType(
       data.refreshToken,
       TokenType.REFRESH,
@@ -134,88 +320,16 @@ export class AuthService {
     );
 
     if (!userEntity) {
-      throw new Error('User not found');
+      this.logger.warn('User not found during token refresh', {
+        userId: tokenEntity.userId,
+      });
+      throw new AuthError(AuthErrorMessage.INVALID_REFRESH_TOKEN);
     }
 
     return this.auth(userEntity);
   }
 
-  public async auth(
-    userEntity: Web2UserEntity | OperatorUserEntity | UserEntity,
-  ): Promise<SuccessAuthDto> {
-    const refreshTokenEntity =
-      await this.tokenRepository.findOneByUserIdAndType(
-        userEntity.id,
-        TokenType.REFRESH,
-      );
-
-    const operatorAddress = this.web3ConfigService.operatorAddress;
-
-    let status = userEntity.status.toString();
-    if (userEntity.role === UserRole.OPERATOR && userEntity.evmAddress) {
-      let operatorStatus: string | undefined;
-      try {
-        operatorStatus = await KVStoreUtils.get(
-          this.web3ConfigService.reputationNetworkChainId,
-          operatorAddress,
-          userEntity.evmAddress.toLowerCase(),
-        );
-      } catch {}
-
-      if (operatorStatus && operatorStatus !== '') {
-        status = operatorStatus;
-      }
-    }
-
-    const payload: any = {
-      email: userEntity.email,
-      status,
-      userId: userEntity.id,
-      wallet_address: userEntity.evmAddress,
-      role: userEntity.role,
-      kyc_status: userEntity.kyc?.status,
-      nda_signed:
-        userEntity.ndaSignedUrl === this.ndaConfigService.latestNdaUrl,
-      reputation_network: operatorAddress,
-      qualifications: userEntity.userQualifications
-        ? userEntity.userQualifications.map(
-            (userQualification) => userQualification.qualification.reference,
-          )
-        : [],
-    };
-
-    // TODO: load sitekeys from repository instead of user entity in request
-    if (userEntity.siteKeys && userEntity.siteKeys.length > 0) {
-      const existingHcaptchaSiteKey = userEntity.siteKeys.find(
-        (key) => key.type === SiteKeyType.HCAPTCHA,
-      );
-      if (existingHcaptchaSiteKey) {
-        payload.site_key = existingHcaptchaSiteKey.siteKey;
-      }
-    }
-
-    const accessToken = await this.jwtService.signAsync(payload, {
-      expiresIn: this.authConfigService.accessTokenExpiresIn,
-    });
-
-    if (refreshTokenEntity) {
-      await this.tokenRepository.deleteOne(refreshTokenEntity);
-    }
-
-    const newRefreshTokenEntity = new TokenEntity();
-    newRefreshTokenEntity.userId = userEntity.id;
-    newRefreshTokenEntity.type = TokenType.REFRESH;
-    const date = new Date();
-    newRefreshTokenEntity.expiresAt = new Date(
-      date.getTime() + this.authConfigService.refreshTokenExpiresIn * 1000,
-    );
-
-    await this.tokenRepository.createUnique(newRefreshTokenEntity);
-
-    return { accessToken, refreshToken: newRefreshTokenEntity.uuid };
-  }
-
-  public async forgotPassword(data: ForgotPasswordDto): Promise<void> {
+  async forgotPassword(data: ForgotPasswordDto): Promise<void> {
     const userEntity = await this.userRepository.findOneByEmail(data.email);
 
     if (!userEntity) {
@@ -245,10 +359,15 @@ export class AuthService {
     });
   }
 
-  public async restorePassword(data: RestorePasswordDto): Promise<void> {
+  async restorePassword(data: RestorePasswordDto): Promise<void> {
     const tokenEntity = await this.tokenRepository.findOneByUuidAndType(
       data.token,
       TokenType.PASSWORD,
+      {
+        relations: {
+          user: true,
+        },
+      },
     );
 
     if (!tokenEntity) {
@@ -259,19 +378,26 @@ export class AuthService {
       throw new AuthError(AuthErrorMessage.REFRESH_TOKEN_EXPIRED);
     }
 
-    const userEntity = await this.userService.updatePassword(
+    const hashedPassword = securityUtils.hashPassword(data.password);
+
+    const isPasswordChanged = await this.userRepository.updateOneById(
       tokenEntity.userId,
-      data.password,
-    );
-    await this.emailService.sendEmail(
-      userEntity.email,
-      EmailAction.PASSWORD_CHANGED,
+      {
+        password: hashedPassword,
+      },
     );
 
-    await this.tokenRepository.deleteOne(tokenEntity);
+    if (isPasswordChanged) {
+      await this.emailService.sendEmail(
+        tokenEntity.user?.email as string,
+        EmailAction.PASSWORD_CHANGED,
+      );
+
+      await this.tokenRepository.deleteOne(tokenEntity);
+    }
   }
 
-  public async emailVerification(data: VerifyEmailDto): Promise<void> {
+  async emailVerification(data: VerifyEmailDto): Promise<void> {
     const tokenEntity = await this.tokenRepository.findOneByUuidAndType(
       data.token,
       TokenType.EMAIL,
@@ -290,16 +416,16 @@ export class AuthService {
     });
   }
 
-  public async resendEmailVerification(
+  async resendEmailVerification(
+    user: Web2UserEntity,
     data: ResendVerificationEmailDto,
   ): Promise<void> {
-    const userEntity = await this.userRepository.findOneByEmail(data.email);
-    if (!userEntity || userEntity.status !== UserStatus.PENDING) {
+    if (user.status !== UserStatus.PENDING) {
       return;
     }
 
     const existingToken = await this.tokenRepository.findOneByUserIdAndType(
-      userEntity.id,
+      user.id,
       TokenType.EMAIL,
     );
 
@@ -309,7 +435,7 @@ export class AuthService {
 
     const tokenEntity = new TokenEntity();
     tokenEntity.type = TokenType.EMAIL;
-    tokenEntity.userId = userEntity.id;
+    tokenEntity.userId = user.id;
     const date = new Date();
     tokenEntity.expiresAt = new Date(
       date.getTime() + this.authConfigService.verifyEmailTokenExpiresIn * 1000,
@@ -319,115 +445,5 @@ export class AuthService {
     await this.emailService.sendEmail(data.email, EmailAction.SIGNUP, {
       url: `${this.serverConfigService.feURL}/verify?token=${tokenEntity.uuid}`,
     });
-  }
-
-  public async web3Signup(data: Web3SignUpDto): Promise<SuccessAuthDto> {
-    const preSignUpData = prepareSignatureBody({
-      from: data.address,
-      to: this.web3ConfigService.operatorAddress,
-      contents: SignatureType.SIGNUP,
-    });
-
-    const verified = verifySignature(preSignUpData, data.signature, [
-      data.address,
-    ]);
-
-    if (!verified) {
-      throw new AuthError(AuthErrorMessage.INVALID_WEB3_SIGNATURE);
-    }
-
-    const chainId = this.web3ConfigService.reputationNetworkChainId;
-
-    const signer = this.web3Service.getSigner(chainId);
-    const kvstore = await KVStoreClient.build(signer);
-
-    let role = '';
-    try {
-      role = await KVStoreUtils.get(chainId, data.address, KVStoreKeys.role);
-    } catch {}
-
-    const validRolesValue = [
-      Role.JobLauncher,
-      Role.ExchangeOracle,
-      Role.RecordingOracle,
-    ].map((role) => role.toLowerCase());
-
-    if (!validRolesValue.includes(role.toLowerCase())) {
-      throw new InvalidOperatorRoleError(role);
-    }
-
-    let fee = '';
-    try {
-      fee = await KVStoreUtils.get(chainId, data.address, KVStoreKeys.fee);
-    } catch {}
-
-    if (!fee) {
-      throw new InvalidOperatorFeeError(fee);
-    }
-
-    let url = '';
-    try {
-      url = await KVStoreUtils.get(chainId, data.address, KVStoreKeys.url);
-    } catch {}
-
-    if (!url) {
-      throw new InvalidOperatorUrlError(url);
-    }
-
-    let jobTypes = '';
-    try {
-      jobTypes = await KVStoreUtils.get(
-        chainId,
-        data.address,
-        KVStoreKeys.jobTypes,
-      );
-    } catch {}
-
-    if (!jobTypes) {
-      throw new InvalidOperatorJobTypesError(jobTypes);
-    }
-
-    const user = await this.userRepository.findOneByAddress(data.address);
-
-    if (user) {
-      throw new DuplicatedUserAddressError(data.address);
-    }
-
-    const userEntity = await this.userService.createOperatorUser(data.address);
-
-    /**
-     * TODO: revisit if we want to make it active by default
-     * since we have `enableOperator` functionality and
-     * they might not have enough tokens, which should not impact signup
-     */
-    await kvstore.set(data.address.toLowerCase(), OperatorStatus.ACTIVE);
-
-    return this.auth(userEntity);
-  }
-
-  public async web3Signin(data: Web3SignInDto): Promise<SuccessAuthDto> {
-    const userEntity = await this.userService.findOperatorUser(data.address);
-
-    if (!userEntity) {
-      throw new AuthError(AuthErrorMessage.INVALID_ADDRESS);
-    }
-
-    const preSigninData = prepareSignatureBody({
-      from: data.address,
-      to: this.web3ConfigService.operatorAddress,
-      contents: SignatureType.SIGNIN,
-      nonce: userEntity.nonce,
-    });
-    const verified = verifySignature(preSigninData, data.signature, [
-      data.address,
-    ]);
-
-    if (!verified) {
-      throw new AuthError(AuthErrorMessage.INVALID_WEB3_SIGNATURE);
-    }
-
-    await this.userService.updateNonce(userEntity);
-
-    return this.auth(userEntity);
   }
 }
